requirements.txt
pillow>=10.0.0
screeninfo>=0.8.1
